requirements.txt
flask>=2.3,<3.0
pillow>=10.0.0
numpy>=1.23.5,<1.25
tensorflow>=2.14,<2.15


